final draft for demo
will have to report overflow as bug
